Arel::Nodes::Infer
