+++ lib/arel/extensions.rb
@@ -49,6 +49,7 @@ require 'arel/extensions/select_statement'
 require 'arel/extensions/insert_statement'
 require 'arel/extensions/default_values'
 require 'arel/extensions/conflict'
+require 'arel/extensions/infer'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
